# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/workouts/exercises": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "muscleGroup",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MuscleGroupEnum"
                }
              ]
            }
          },
          {
            "name": "metricType",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ExerciseMetricEnum"
                }
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exercises"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateExerciseRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateExerciseRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateExerciseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/exercises/{id}": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateExerciseRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateExerciseRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateExerciseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workouts/exercises/{id}/archived": {
      "patch": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetExerciseArchivedRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetExerciseArchivedRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SetExerciseArchivedRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workouts/sessions/{id}/exercises/{entryId}": {
      "delete": {
        "tags": [
          "WorkoutSessions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSessionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSessionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateExerciseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "metricType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExerciseMetricEnum"
              }
            ]
          },
          "muscleGroup": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MuscleGroupEnum"
              }
            ]
          },
          "equipment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EquipmentEnum"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EquipmentEnum": {
        "enum": [
          "None",
          "Barbell",
          "Dumbbell",
          "Kettlebell",
          "Machine",
          "Cable",
          "ResistanceBand",
          "Other"
        ],
        "type": "string"
      },
      "ExerciseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "metricType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExerciseMetricEnum"
              }
            ]
          },
          "muscleGroup": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MuscleGroupEnum"
              }
            ]
          },
          "equipment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EquipmentEnum"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean"
          },
          "isArchived": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ExerciseMetricEnum": {
        "enum": [
          "Reps",
          "RepsAndWeight",
          "Time",
          "Distance",
          "DistanceAndTime"
        ],
        "type": "string"
      },
...
      "MuscleGroupEnum": {
        "enum": [
          "Other",
          "Chest",
          "Back",
          "Shoulders",
          "Biceps",
          "Triceps",
          "Forearms",
          "Abs",
          "Glutes",
          "Quadriceps",
          "Hamstrings",
          "Calves",
          "FullBody",
          "Cardio"
        ],
        "type": "string"
      },
...
      "SetExerciseArchivedRequest": {
        "type": "object",
        "properties": {
          "isArchived": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateExerciseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "metricType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExerciseMetricEnum"
              }
            ]
          },
          "muscleGroup": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MuscleGroupEnum"
              }
            ]
          },
          "equipment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EquipmentEnum"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WorkoutSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "routineId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "performedOn": {
            "type": "string",
            "format": "date"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutSessionStatusEnum"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "totals": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutSessionTotalsDto"
              }
            ],
            "nullable": true
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSessionExerciseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutSessionExerciseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "exerciseName": {
            "type": "string",
            "nullable": true
          },
          "metricType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExerciseMetricEnum"
              }
            ]
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "targetSets": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetReps": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetWeight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "targetDurationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetDistance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "restSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutSessionStatusEnum": {
        "enum": [
          "InProgress",
          "Completed",
          "Abandoned"
        ],
        "type": "string"
      },
...
      "WorkoutSessionTotalsDto": {
        "type": "object",
        "properties": {
          "exerciseCount": {
            "type": "integer",
            "format": "int32"
          },
          "setCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalReps": {
            "type": "integer",
            "format": "int32"
          },
          "totalVolume": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "WorkoutSetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "setNumber": {
            "type": "integer",
            "format": "int32"
          },
          "reps": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "distance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "rpe": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "setType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkoutSetTypeEnum"
              }
            ]
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "estimatedOneRepMax": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutSetTypeEnum": {
        "enum": [
          "Normal",
          "Warmup",
          "DropSet",
          "FailureSet"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter ONLY token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "x-time-zone": {
        "type": "apiKey",
        "description": "Specify the user's timezone in IANA format (e.g., 'America/New_York')",
        "name": "x-time-zone",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    },
    {
      "x-time-zone": [ ]
    }
  ]
}